+++ lib/concurrent_rails/promises.rb
@@ -2,11 +2,17 @@
 
 module ConcurrentRails
   class Promises
-    include Concurrent::Promises::FactoryMethods
     extend Forwardable
+    include Concurrent::Promises::FactoryMethods
 
-    def self.future(*args, &task)
-      new.run_on_rails(*args, &task)
+    class << self
+      def future(*args, &task)
+        future_on(:fast, *args, &task)
+      end
+
+      def future_on(executor, *args, &task)
+        new.with_rails(executor, *args, &task)
+      end
     end
 
     def then(*args, &task)
@@ -18,18 +24,18 @@ module ConcurrentRails
     end
 
     %i[value value!].each do |method_name|
-      define_method method_name do
+      define_method method_name do |timeout = nil, timeout_value = nil|
         Rails.application.executor.wrap do
           ActiveSupport::Dependencies.interlock.permit_concurrent_loads do
-            future_instance.__send__(method_name)
+            future_instance.__send__(method_name, timeout, timeout_value)
           end
         end
       end
     end
 
-    def run_on_rails(*args, &task)
+    def with_rails(executor, *args, &task)
       @future_instance = Rails.application.executor.wrap do
-        future_on(default_executor, *args, &task)
+        future_on(executor, *args, &task)
       end
 
       self